Move MemberInfoMetadataComparer to own file
Also renames SchemaGenerationContext.AsDeclaredMemberInfoComparer to DeclarationOrderComparer
